backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "fastapi-cache2>=0.2.2",
    "fastapi[uvicorn]>=0.123.5",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt[crypto]>=2.10.1",
    "redis>=7.1.0",
    "sqlalchemy[async]>=2.0.44",
    "uvicorn>=0.38.0",
]
